+++ src/ReleaseNotes.md
@@ -1,3 +1,11 @@
+# 0.2.0 (2018-02-27)
+
+## 新增功能
+
+1. `getFormErrors` 添加 `options.multi` 用于指示是否搜集同一个字段的多个错误信息 (默认为`false`).
+2. `getFormErrors` 添加 `options.falsyKey` 用于指示是否搜集值为`falsy`的错误 (默认为`false`).
+3. `getFormErrors` 现在会尝试搜集 `error.name` 格式的错误信息.
+
 # 0.1.0 (2017-11-01)
 
 ## 初版